changed isModelReady and isModelGenerating to isModel and isReady for alignment with naming semantics
