+++ docs/docs/guides/running-llms.md
@@ -23,7 +23,7 @@ const llama = useLLM({
 });
 ```
 
-The code snippet above fetches the model from the specified URL, loads it into memory, and returns an object with various methods and properties for controlling the model. You can monitor the loading progress by checking the `llama.downloadProgress` and `llama.isModelReady` property, and if anything goes wrong, the `llama.error` property will contain the error message.
+The code snippet above fetches the model from the specified URL, loads it into memory, and returns an object with various methods and properties for controlling the model. You can monitor the loading progress by checking the `llama.downloadProgress` and `llama.isReady` property, and if anything goes wrong, the `llama.error` property will contain the error message.
 
 :::danger[Danger]
 Lower-end devices might not be able to fit LLMs into memory. We recommend using quantized models to reduce the memory footprint.
@@ -50,9 +50,9 @@ Given computational constraints, our architecture is designed to support only on
 | `generate`          | `(input: string) => Promise<void>` | Function to start generating a response with the given input string.                                            |
 | `response`          | `string`                           | State of the generated response. This field is updated with each token generated by the model                   |
 | `error`             | <code>string &#124; null</code>    | Contains the error message if the model failed to load                                                          |
-| `isModelGenerating` | `boolean`                          | Indicates whether the model is currently generating a response                                                  |
+| `isGenerating` | `boolean`                          | Indicates whether the model is currently generating a response                                                  |
 | `interrupt`         | `() => void`                       | Function to interrupt the current inference                                                                     |
-| `isModelReady`      | `boolean`                          | Indicates whether the model is ready                                                                            |
+| `isReady`      | `boolean`                          | Indicates whether the model is ready                                                                            |
 | `downloadProgress`  | `number`                           | Represents the download progress as a value between 0 and 1, indicating the extent of the model file retrieval. |
 
 ### Loading models
