+++ examples/computer-vision/screens/ClassificationScreen.tsx
@@ -44,10 +44,10 @@ export const ClassificationScreen = ({
     }
   };
 
-  if (!model.isModelReady) {
+  if (!model.isReady) {
     return (
       <Spinner
-        visible={!model.isModelReady}
+        visible={!model.isReady}
         textContent={`Loading the model...`}
       />
     );
