+++ examples/llama/screens/ChatScreen.tsx
@@ -31,10 +31,10 @@ export default function ChatScreen() {
   });
   const textInputRef = useRef<TextInput>(null);
   useEffect(() => {
-    if (llama.response && !llama.isModelGenerating) {
+    if (llama.response && !llama.isGenerating) {
       appendToMessageHistory(llama.response, 'ai');
     }
-  }, [llama.response, llama.isModelGenerating]);
+  }, [llama.response, llama.isGenerating]);
 
   const appendToMessageHistory = (input: string, role: SenderType) => {
     setChatHistory((prevHistory) => [
@@ -54,9 +54,9 @@ export default function ChatScreen() {
     }
   };
 
-  return !llama.isModelReady ? (
+  return !llama.isReady ? (
     <Spinner
-      visible={!llama.isModelReady}
+      visible={!llama.isReady}
       textContent={`Loading the model ${(llama.downloadProgress * 100).toFixed(0)} %`}
     />
   ) : (
@@ -76,7 +76,7 @@ export default function ChatScreen() {
               <Messages
                 chatHistory={chatHistory}
                 llmResponse={llama.response}
-                isModelGenerating={llama.isModelGenerating}
+                isGenerating={llama.isGenerating}
               />
             </View>
           ) : (
@@ -108,13 +108,13 @@ export default function ChatScreen() {
               <TouchableOpacity
                 style={styles.sendChatTouchable}
                 onPress={async () =>
-                  !llama.isModelGenerating && (await sendMessage())
+                  !llama.isGenerating && (await sendMessage())
                 }
               >
                 <SendIcon height={24} width={24} padding={4} margin={8} />
               </TouchableOpacity>
             )}
-            {llama.isModelGenerating && (
+            {llama.isGenerating && (
               <TouchableOpacity
                 style={styles.sendChatTouchable}
                 onPress={llama.interrupt}
