+++ src/index.tsx
@@ -4,4 +4,3 @@ export * from './constants/modelUrls';
 export * from './models/Classification';
 export * from './models/ObjectDetection';
 export * from './models/StyleTransfer';
-export * from './types/object_detection_types';
