+++ src/models/Classification.ts
@@ -8,8 +8,8 @@ interface Props {
 
 interface ClassificationModule {
   error: string | null;
-  isModelReady: boolean;
-  isModelGenerating: boolean;
+  isReady: boolean;
+  isGenerating: boolean;
   forward: (input: string) => Promise<{ [category: string]: number }>;
 }
 
@@ -17,10 +17,10 @@ export const useClassification = ({
   modelSource,
 }: Props): ClassificationModule => {
   const [module, _] = useState(() => new _ClassificationModule());
-  const { error, isModelReady, isModelGenerating, forward } = useModule({
+  const { error, isReady, isGenerating, forwardImage: forward } = useModule({
     modelSource,
     module,
   });
 
-  return { error, isModelReady, isModelGenerating, forward };
+  return { error, isReady, isGenerating, forward };
 };
